Keep set of gathered resources
We want to add addons collecting related cluster resources. This
introduced the problem of collecting the same cluster resource
concurrently, for example storageclass used by many pvcs.

To allow gathering every resource once, we keep now a set of gathered
resources. We update the set before dumping or inspecting resources and
skip the resource if it was already gathered.

Having a set of resources we can remove the atomic counter and use the
length of the set.

Having a key for gathered resources, we can simplify logging of gathered
resources and errors using the key.

+++ cmd/local.go
@@ -12,7 +12,7 @@ import (
 )
 
 type result struct {
-	Count int32
+	Count int
 	Err   error
 }
 
@@ -72,7 +72,7 @@ func localGather(clusters []*clusterConfig) {
 	wg.Wait()
 	close(results)
 
-	count := int32(0)
+	count := 0
 
 	for r := range results {
 		if r.Err != nil {
